+++ Modules/Common/Targets/Model/Sources/Domain/Input/CreatePollCommentsRequestInput.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public struct CreatePollCommentRequestInput: Encodable {
+    public let content: String
+
+    public init(content: String) {
+        self.content = content
+    }
+}
