+++ Modules/Common/Targets/Model/Sources/Domain/Response/IdResponseString.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public struct IdResponseString: Decodable {
+    public let id: String
+}
